+++ hubmap/frontend/src/ui-components/SearchInput.js
