+++ hubmap/frontend/src/ui-components/index.js
@@ -2,8 +2,8 @@
  * This exports all required components for browse data page.
  */
 
-export { default as CustomSearchInput } from './CustomSearchInput';
+export { default as SearchInput } from './SearchInput';
 export { default as ControlledChipInput } from './ControlledChipInput';
-export { default as CustomSelectMenu } from './CustomSelectMenu';
-export { default as SelectedBrowseOptionsSummary } from './SelectedBrowseOptionsSummary';
+export { default as BaseSelectDropdown } from './BaseSelectDropdown';
+export { default as BrowseSummary } from './BrowseSummary';
 export { default as ReactAutosuggestExample } from './AutoSuggest';
